+++ Orleans.Streams.Kafka/Core/KafkaBatchContainer.cs
@@ -1,7 +1,6 @@
 ﻿using Confluent.Kafka;
 using Orleans.Providers.Streams.Common;
 using Orleans.Runtime;
-using Orleans.Streams.Utils.MessageTracking;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -9,12 +8,11 @@ using System.Linq;
 namespace Orleans.Streams.Kafka.Core
 {
 	[Serializable]
-	public class KafkaBatchContainer : ITraceablebleBatch, IComparable<KafkaBatchContainer>
+	public class KafkaBatchContainer : IBatchContainer, IComparable<KafkaBatchContainer>
 	{
 		private readonly Dictionary<string, object> _requestContext;
 
 		[NonSerialized] internal TopicPartitionOffset TopicPartitionOffSet;
-
 
 		protected List<object> Events { get; set; }
 
@@ -23,8 +21,6 @@ namespace Orleans.Streams.Kafka.Core
 		public string StreamNamespace { get; }
 
 		public StreamSequenceToken SequenceToken { get; internal set; }
-
-		public List<object> RawEvents => Events;
 
 		public KafkaBatchContainer(
 			Guid streamGuid,