+++ Samples/TestSilo/Program.cs
@@ -22,6 +22,13 @@ namespace TestSilo
 			const int gatewayPort = 30000;
 			var siloAddress = IPAddress.Loopback;
 
+			var brokers = new List<string>
+			{
+				"[host name]:39000",
+				"[host name]:39001",
+				"[host name]:39002"
+			};
+
 			var builder = new SiloHostBuilder()
 				.Configure<ClusterOptions>(options =>
 				{
@@ -38,10 +45,12 @@ namespace TestSilo
 				.AddKafka("KafkaProvider")
 				.WithOptions(options =>
 				{
-					options.BrokerList = new List<string> { "localhost:9092" };
+					options.BrokerList = brokers.ToArray();
 					options.ConsumerGroupId = "TestGroup";
-					options.Topics = new List<TopicConfig> { new TopicConfig { Name = "gossip-testing" } };
 					options.MessageTrackingEnabled = true;
+					options.AddTopic("sucrose-test");
+					options.AddTopic("sucrose-auto", new TopicCreationConfig { AutoCreate = true, Partitions = 2, ReplicationFactor = 1 });
+					options.AddTopic("sucrose-auto2", new TopicCreationConfig { AutoCreate = true, Partitions = 3, ReplicationFactor = 1 });
 				})
 				.AddLoggingTracker()
 				.Build();